+++ spec/lib/karafka/web/ui/controllers/status_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  subject(:app) { Karafka::Web::Ui::App }
+
+  describe '#show' do
+    before { get 'status' }
+
+    it { expect(last_response).to be_ok }
+  end
+end