+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - [Fix] Make sure, that when re-using same Karafka Web-UI topics as a different up, all states and reports are not recomputed back.
 - [Fix] Fix headers size inconsistency between Health and Routing.
 - [Fix] Fix invalid padding on status page.
+- [Fix] Fix a case wher root mounted Karafka Web-UI would not work.
 - [Refactor] Limit usage of UI models for data intense computation to speed up states materialization under load.
 - [Refactor] Reorganize pagination engine to support offset based pagination.
 - [Maintenance] Require `karafka` `2.1.8` due to fixes in the Iterator API and routing API extensions.